CSLR/stochastic-cslr/requirements.txt
pandas
xmltodict
einops
torch>=1.8.0
torchvision>=0.8.1
zouqi==1.0.5
opencv-python>=4.1.2
PyYAML>=5.3
treelib
natsort
imageio
pandarallel
arpa
efficientnet_pytorch
decord
Pillow
python-Levenshtein
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
scipy>=1.4.1
tensorboard>=2.2
tqdm>=4.41.0